On branch MF_only_exc_inh_sep 	new file:   neural_modelling/src/meanfield/NOTE_a_faire.txt 	modified:   neural_modelling/src/meanfield/c_main_neurons.c 	modified:   neural_modelling/src/meanfield/implementations/meanfield_impl_standard.h 	modified:   neural_modelling/src/meanfield/input_types/input_type_conductance.h 	modified:   neural_modelling/src/meanfield/meanfield.c 	modified:   neural_modelling/src/meanfield/meanfield_recording.h 	modified:   neural_modelling/src/meanfield/models/meanfield_model_impl.c 	modified:   neural_modelling/src/meanfield/plasticity/synapse_dynamics_static_impl.c 	modified:   neural_modelling/src/meanfield/send_spike.h 	modified:   neural_modelling/src/meanfield/synapse_types/synapse_types_exponential_impl.h 	modified:   neural_modelling/src/meanfield/synapses.c
Do some change on this branch in order to find a F... way to connect MFs together and sens the number directly
without passing by send_spike, buffer and Co.
I will extract this branch in order to save it and create a clean branch, removing all log that are not necessary and
all old code versions. I hope this give me an better view.